pyproject.toml
[tool.poetry]
name = "leesah"
version = "1.0.0"
description = "Leesah er et Pythonbibliotek for å spille det hendelsedrevet applikasjonsutviklingspillet Leesah Game"
authors = ["NAV"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
schema = "^0.7.5"
confluent-kafka = "^2.3.0"
PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
